requirements.txt
--extra-index-url https://download.pytorch.org/whl/torch_stable.html
protobuf==3.20.3
qtconsole==5.4.2
aiohttp==3.8.4
numpy==1.24.2
torch==2.3.1+cu121 -f https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.3.1+cu121 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.18.1+cu121 -f https://download.pytorch.org/whl/torch_stable.html
lightning==2.2.5
tensorboard==2.12.2
tensorboardx==2.6
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
setuptools==67.6.1
matplotlib==3.6.3
networkx==3.1
requests==2.28.2
torchmetrics==0.11.0
Pillow==9.4.0
ansi2html==1.8.0
Werkzeug==2.2.3
pycryptodome==3.16.0
Sphinx==6.1.3
sphinx-autoapi==2.0.0
psutil==5.9.4
wandb==0.13.9
rich==13.3.4
seaborn==0.12.2
scikit-learn==1.2.2
scikit-image==0.20.0
pandas==1.5.2
timm==0.6.13
gunicorn==20.1.0
nvidia-ml-py==11.525.84
scikit-image
